more powershell implementions

+++ powershell/scripts/sharedmailboxes/remove.ps1
@@ -1,8 +1,13 @@
 # Connector: Exchange
 # Commands: Remove-Mailbox 
 param (
-    $ExchangeObjectId
+    [Parameter(Mandatory = $true)]
+    [string]$ExchangeObjectId
 )
-
+$mb = Get-MailBox -Identity $ExchangeObjectId
+if ($mb -eq $null) {
+    write-output "Mailbox $ExchangeObjectId not found"
+    exit 1
+}
 Remove-Mailbox -Identity $ExchangeObjectId -Confirm:$false
 write-output "done"
